requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord.py @ git+https://github.com/Rapptz/discord.py@868476c99963e72aa529ed6203f86ef529a2c1f1
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
typing_extensions==4.2.0
yarl==1.7.2
